requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
sqlalchemy[asyncio]>=2.0.25
alembic>=1.13.0
asyncpg>=0.29.0
pgvector>=0.2.5
psycopg2-binary>=2.9.9
pydantic>=2.5.0
pydantic-settings>=2.1.0
anthropic>=0.40.0
openai>=1.12.0
pdfplumber>=0.11.0
python-docx>=1.1.0
markdown>=3.5.0
atlassian-python-api>=3.41.0
httpx>=0.27.0
